+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Component definitions for Pip.Services in Go Changelog
 
+
+## <a name="1.3.0"></a> 1.4.0 (2021-10-24) 
+Added state management components
+### Features
+* **state** Added IStateStore interface and StateValue struct
+* **state** Added NullStateStore struct
+* **state** Added MemoryStateStore struct
+* **state** Added DefaultStateStoreFactory struct
+* **config** Extended IConfigReader interface (added *AddChangeListener* and *RemoveChangeListener* methods)
+
 ## <a name="1.2.0"></a> 1.2.0 (2021-04-23) 
 
 ### Features